Revert "msm: kgsl: Affine IRQ and worker kthread to the big CPU cluster"
This reverts commit 8088c690d2d9a9016f06f400eebfc2623b68d4f4.

This is wrong. This causes 3 IRQs to be affined to the big CPU cluster,
not just the primary kgsl_3d0_irq one. As a result, the perf crit API
thinks that the 2 extra IRQs are critical and will balance them despite
them being rarely used (kgsl_hfi_irq and kgsl_gmu_irq).

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: VoidDev-0 <dev.avax@protonmail.com>